+++ springsecuritybasic/src/main/java/com/anuj/test/springsecuritybasic/repository/AccountTransactionsRepository.java
@@ -7,5 +7,5 @@ import java.util.List;
 
 @Repository
 public interface AccountTransactionsRepository extends CrudRepository<AccountTransactions,Long> {
-    List<AccountTransactions> findByCustomerIdOrderByTransactionDtDesc(int customerId);
+    List<AccountTransactions> findByCustomerIdOrderByTransactionDtDesc(long customerId);
 }
