+++ springsecuritybasic/src/main/java/com/anuj/test/springsecuritybasic/repository/LoanRepository.java
@@ -7,6 +7,6 @@ import java.util.List;
 
 @Repository
 public interface LoanRepository extends CrudRepository<Loans,Long> {
-    List<Loans> findByCustomerIdOrderByStartDtDesc(int customerId);
+    List<Loans> findByCustomerIdOrderByStartDtDesc(long customerId);
 
 }
